requirements.txt
google-genai
fastapi
uvicorn
langgraph
langchain
python-dotenv
pydantic
python-multipart
aiofiles
